@@ -1,8 +1,10 @@
 module github.com/k4cg/k4cgbot
 
-go 1.15
+go 1.22
 
 require (
 	github.com/mb-14/gomarkov v0.0.0-20200824165802-565ae3f3c79f
 	gopkg.in/tucnak/telebot.v2 v2.3.4
 )
+
+require github.com/pkg/errors v0.8.1 // indirect
